UserHome/requirements.txt
aiofiles==0.3.2
gunicorn==19.8.0
httptools==0.0.11
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.1
motor==1.2.1
pymongo==3.6.1
python-dateutil==2.7.2
pytz==2018.4
sanic==0.7.0
sanic-base-extension==0.1.1
sanic-jinja2==0.7.0
sanic-mongodb-extension==0.1.1
six==1.11.0
ujson==1.35
umongo==0.15.0
uvloop==0.9.1
websockets==4.0.1
